feat(DEV-11277): add abstract `MoniteAppElementBase` class for future usage in other web components
